atualização no controller user

+++ src/controller/UserController.js
@@ -18,17 +18,18 @@ module.exports = {
         try {
             const { name, email } = req.body;
 
-            const user = await User.findOne({email})
+            let user = await User.findOne({where: {email}})
 
             if(user) {
                 return res.json({massage: "existing user"})
+            } else {
+                user = await User.create({ name, email });
             }
 
-            user = await User.create({ name, email });
-
             return res.json({ user });
-        } catch (error) {
-            return res.json({ error: 'failed to create user' });
+            
+        } catch (err) {
+            return res.json({ error: 'failed to create user' + err});
         }
     },
     async update(req, res) {
@@ -36,24 +37,28 @@ module.exports = {
             const { id } = req.params;
             const { name, email } = req.body;
 
-            const user = await User.update({ name, email }, { where: { id } });
+            let user = await User.update({ name, email }, { where: { id } });
 
-            return res.json({ user });
-        } catch (error) {
-            return res.json({ error: 'failed to update user' });
+            return res.json(user);
+        } catch (err) {
+            return res.json({ error: 'failed to update user ' + err});
         }
     },
     async delete(req, res) {
         try {
             const { id } = req.params;
 
-            const user = await User.
+            const user = await User.findByPk(id)
 
-            User.destroy({ where: { id } });
+            if(!user) {
+                return res.status(400).json({ error: 'user not found' })
+            }
 
-            return res.json();
+            user = await User.destroy({ where: { id } });
+
+            return res.json({ massage: 'success to delete user' });
         } catch (error) {
-            return res.json({ error: 'failed to delete user' });
+            return res.json({ error: 'failed to delete user', error });
         }
 
     },
